# docs/openapi/panel.yaml
openapi: 3.0.3
info:
  title: Lotty AB Platform — Panel API
  description: Control plane API.
  version: 0.0.0_dev
servers:
  - url: http://localhost:8080/api/panel/v1
    description: local development

# JWT bearer auth; endpoints marked public below are excluded
security:
  - bearerAuth: []
...
paths:
...
  /me:
    get:
      tags: [users]
      summary: Get the authenticated user's own profile
      operationId: getMe
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/avatar:
    post:
      tags: [users]
      summary: Upload or delete the authenticated user's avatar
      description: |
        Upload a new avatar by sending a file in the `avatar` field.
        Delete the avatar by sending an empty `avatar` field or omitting it entirely.
      operationId: uploadMyAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
              required: [avatar]
      responses:
        "200":
          description: Avatar uploaded or deleted, user profile returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /users:
    post:
      tags: [users]
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [users]
      summary: List users
      operationId: listUsers
      parameters:
        - name: limit
          in: query
          description: Maximum number of users to return (1-100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of users to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          format: int64
          minimum: 1

    get:
      tags: [users]
      summary: Get a user by ID
      operationId: getUser
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      tags: [users]
      summary: Partially update a user
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [users]
      summary: Soft-delete a user
      operationId: deleteUser
      responses:
        "204":
          description: User deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /users/{id}/avatar:
    parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          format: int64
          minimum: 1

    post:
      tags: [users]
      summary: Upload or delete a user's avatar (admin only)
      description: |
        Upload a new avatar by sending a file in the `avatar` field.
        Delete the avatar by sending an empty `avatar` field or omitting it entirely.
      operationId: uploadUserAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
              required: [avatar]
      responses:
        "200":
          description: Avatar uploaded or deleted, user profile returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /flags:
    post:
      tags: [flags]
      summary: Create a feature flag
      operationId: createFlag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlagRequest"
      responses:
        "200":
          description: Flag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [flags]
      summary: List feature flags
      operationId: listFlags
      parameters:
        - name: limit
          in: query
          description: Maximum number of flags to return (1-100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of flags to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /flags/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Flag ID
        schema:
          type: integer
          format: int64
          minimum: 1

    get:
      tags: [flags]
      summary: Get a flag by ID
      operationId: getFlag
      responses:
        "200":
          description: Flag found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      tags: [flags]
      summary: Partially update a flag
      operationId: updateFlag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFlagRequest"
      responses:
        "200":
          description: Flag updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [flags]
      summary: Soft-delete a flag
      operationId: deleteFlag
      responses:
        "204":
          description: Flag deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: BAD_REQUEST
              message: "Key: 'CreateUserRequest.Email' Error:Field validation for 'Email' failed on the 'email' tag"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: NOT_FOUND
              message: user not found
    Conflict:
      description: Resource already exists (username or email taken)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: CONFLICT
              message: user already exists
    Forbidden:
      description: Action forbidden (e.g. self-delete, self-role-change, or last-admin protection)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: FORBIDDEN
              message: cannot delete your own account
    Unauthorized:
      description: Authentication required or token invalid/revoked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: missing or malformed token
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: INTERNAL_SERVER_ERROR
              message: internal server error
  schemas:
...
    Role:
      type: string
      enum: [admin, experimenter, approver, viewer]
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          minLength: 3
          maxLength: 64
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        avatar_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, username, email, role, created_at, updated_at]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        role:
          $ref: "#/components/schemas/Role"
      required: [username, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
      minProperties: 1
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/User"
      required: [success, data]
    PaginationMeta:
      type: object
      required: [limit, offset, count, total, has_next]
      properties:
        limit:
          type: integer
          minimum: 1
          example: 20
        offset:
          type: integer
          minimum: 0
          example: 0
        count:
          type: integer
          minimum: 0
          example: 20
        total:
          type: integer
          format: int64
          minimum: 0
          example: 137
        has_next:
          type: boolean
          example: true
    UserListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                [
                  BAD_REQUEST,
                  NOT_FOUND,
                  PAYLOAD_TOO_LARGE,
                  CONFLICT,
                  FORBIDDEN,
                  INTERNAL_SERVER_ERROR,
                ]
            message:
              type: string
          required: [code, message]
      required: [success, error]
...
    FlagType:
      type: string
      enum: [string, number, bool]
    Flag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
          minLength: 3
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 256
        type:
          $ref: "#/components/schemas/FlagType"
        default_value:
          description: JSON value matching the flag type (string, number, or boolean)
          oneOf:
            - type: string
            - type: number
            - type: boolean
        description:
          type: string
          maxLength: 4096
        created_by:
          $ref: "#/components/schemas/User"
        updated_by:
          $ref: "#/components/schemas/User"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        [
          id,
          key,
          name,
          type,
          default_value,
          created_by,
          updated_by,
          created_at,
          updated_at,
        ]
    CreateFlagRequest:
      type: object
      properties:
        key:
          type: string
          minLength: 3
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 256
        type:
          $ref: "#/components/schemas/FlagType"
        default_value:
          description: JSON value matching the flag type
          oneOf:
            - type: string
            - type: number
            - type: boolean
        description:
          type: string
          minLength: 1
          maxLength: 4096
      required: [key, name, type, default_value]
    UpdateFlagRequest:
      type: object
      properties:
        key:
          type: string
          minLength: 3
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 256
        default_value:
          description: JSON value matching the flag type
          oneOf:
            - type: string
            - type: number
            - type: boolean
        description:
          type: string
          minLength: 1
          maxLength: 4096
      minProperties: 1
    FlagResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Flag"
      required: [success, data]
    FlagListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/FlagListItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    FlagListItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
          minLength: 3
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 256
        type:
          $ref: "#/components/schemas/FlagType"
        default_value:
          description: JSON value matching the flag type (string, number, or boolean)
          oneOf:
            - type: string
            - type: number
            - type: boolean
        description:
          type: string
          maxLength: 4096
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, key, name, type, default_value, created_at, updated_at]
...